fix timestamp exceeding max int value

+++ src/main/java/com/strap/sdk/java/StrapReportModel.java
@@ -14,7 +14,7 @@ import java.util.Map;
 public class StrapReportModel {
 
     public String id;
-    public int timestamp;
+    public long timestamp;
     public String date;
     public String type;
     public String range;
